api/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn
psycopg2-binary==2.9.6
django-cors-headers==4.2.0
azure-storage-blob
azure-identity
